backend/requirements.txt
# FastAPI and dependencies
fastapi==0.104.1
uvicorn==0.24.0.post1
starlette==0.27.0

# HTTP client
requests==2.31.0
httpx==0.25.2

# Data processing
numpy==1.24.4
pandas==2.0.3

# Environment variables
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==0.8.11

# Vector database
faiss-cpu==1.7.4

# Embedding models
sentence-transformers==2.2.2

# Type checking
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
alembic==1.12.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pytest==7.4.3
